+++ ClosedXML/Excel/Ranges/XLRangeConsolidationEngine.cs
@@ -117,8 +117,8 @@ namespace ClosedXML.Excel
 
             private readonly IXLWorksheet _worksheet;
             private Dictionary<int, BitArray> _bitMatrix;
-            private int _maxColumn = 0;
-            private int _minColumn = XLHelper.MaxColumnNumber + 1;
+            private short _maxColumn = 0;
+            private short _minColumn = XLHelper.MaxColumnNumber + 1;
 
             #endregion Private Fields
